llamaindex_full/requirements.txt
llama-index>=0.9.0
llama-index-vector-stores-qdrant
llama-index-embeddings-huggingface
llama-index-llms-openai-like
sentence-transformers>=2.2.0
qdrant-client>=1.7.0
python-dotenv